Swagger-UI Structure finished
