+++ nodemon.json
@@ -1,0 +1,3 @@
+{
+  "ext": ".js,.mjs,.ts,.mts,.json,.yaml"
+}
