+++ package.json
@@ -47,7 +47,9 @@
     "mongoose": "^7.0.3",
     "morgan": "^1.10.0",
     "nodemailer": "^6.9.1",
+    "nodemon": "^3.0.1",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.0"
+    "swagger-ui-express": "^5.0.0",
+    "yamljs": "^0.3.0"
   }
 }
